gometalinter: add deadcode, goimports, unconvert
... and fix some warnings from unconvert:

ca/config.go:627:58:warning: unnecessary conversion (unconvert)
manager/allocator/cnmallocator/portallocator.go:410:38:warning: unnecessary conversion (unconvert)
manager/allocator/cnmallocator/portallocator.go:415:50:warning: unnecessary conversion (unconvert)
manager/controlapi/service.go:200:47:warning: unnecessary conversion (unconvert)
manager/controlapi/service.go:210:45:warning: unnecessary conversion (unconvert)
manager/controlapi/service.go:220:35:warning: unnecessary conversion (unconvert)
manager/dispatcher/nodes.go:159:33:warning: unnecessary conversion (unconvert)
manager/state/store/memory.go:692:91:warning: unnecessary conversion (unconvert)

(cherry picked from commit 278edc28b8c4330496c01f39b72c32a054af766d)

+++ manager/state/store/memory.go
@@ -689,7 +689,7 @@ func (tx readTx) findIterators(table string, by By, checkType func(By) error) ([
 		}
 		return []memdb.ResultIterator{it}, nil
 	case bySlot:
-		it, err := tx.memDBTx.Get(table, indexSlot, v.serviceID+"\x00"+strconv.FormatUint(uint64(v.slot), 10))
+		it, err := tx.memDBTx.Get(table, indexSlot, v.serviceID+"\x00"+strconv.FormatUint(v.slot, 10))
 		if err != nil {
 			return nil, err
 		}
